requirements.txt
chatterbot
sklearn
numpy
scipy
speechrecognition
google_speech
pyaml
pyaudio